// backend/src/users/user.dto.ts
import { type BaseSchema } from "../common/dto/base.dto";

export interface IUser extends BaseSchema {
  name: string;
  email: string;
  password: string;
  online: boolean;
  role: "ADMIN" | "USER" | "STAFF";
  refreshToken?: string;
  forgotPasswordToken?: string;
  forgotPasswordTokenExpiry?: Date | "";
}